Resolve comments on PR
