+++ fundamentals/android/auth-sign-in-with-google/Notes/app/src/main/java/com/notes/app/NotesSnackbarManager.kt
@@ -2,12 +2,11 @@ package com.notes.app
 
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asStateFlow
 
 object SnackbarManager {
     private val messages: MutableStateFlow<String?> = MutableStateFlow(null)
     val snackbarMessages: StateFlow<String?>
-        get() = messages.asStateFlow()
+        get() = messages
 
     fun showMessage(message: String) {
         messages.value = message
