+++ fundamentals/android/auth-sign-in-with-google/Notes/app/src/main/java/com/notes/app/screens/account_center/AccountCenterButtons.kt
@@ -10,7 +10,6 @@ import androidx.compose.material.icons.filled.Edit
 import androidx.compose.material.icons.filled.ExitToApp
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Card
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
@@ -24,12 +23,9 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import com.notes.app.R
-import com.notes.app.ui.theme.Purple40
 
 @Composable
 @OptIn(ExperimentalMaterial3Api::class)
@@ -157,29 +153,6 @@ fun RemoveAccountCard(onRemoveAccountClick: () -> Unit) {
                 }
             },
             onDismissRequest = { showRemoveAccDialog = false }
-        )
-    }
-}
-
-@Composable
-fun AuthenticationButton(buttonText: Int, onButtonClick: () -> Unit) {
-    Button(
-        onClick = { onButtonClick() },
-        colors = ButtonDefaults.buttonColors(containerColor = Purple40),
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(16.dp, 0.dp)
-    ) {
-        Icon(
-            painter = painterResource(id = R.drawable.google_g),
-            modifier = Modifier.padding(horizontal = 16.dp),
-            contentDescription = "Google logo"
-        )
-
-        Text(
-            text = stringResource(buttonText),
-            fontSize = 16.sp,
-            modifier = Modifier.padding(0.dp, 6.dp)
         )
     }
 }
